End Of Loop Audio
Added end of loop audio configuration.
Supports in-game sounds for the average user. Limitation of in-game sounds is that they play on the System Sounds channel.
As an alternative so that sounds still play when you have the Master/SystemSounds muted, I added NAudio with the option to enter a custom sound file path. These sounds will play even when you have the prior sound channels muted in game.
(Also sneaking in a Howling Eye path for Questionable future use)