Updated (442) Alexander - The Fist of the Father
Uses Conditions and a bit of AutoMove instead of vnav to make it through the "RNG room", since personally vnav wasn't working for me.

I did a few hundred runs of Fist of the Father for HW Relic Weapons, and this is the final version of the Path that I settled on as I believe it is the most consistent (had no issues with the last ~75 runs with this final iteration). I believe there might still be a _very_ small chance for it to get stuck (if you need to go to the right and steam hits you right as Arm's Length / Surecast falls off), but even in that worst-case scenario, AutoDuty just teleports back to the beginning and tries the Path again, and the chances of steam fucking you up again are extremely small. And even if it does, it'll just port and try again... So it should never get hard-stuck. Like I said, though, I didn't even have this happen across 75 runs or so with this version of the Path, so I digress.